Make SSL domain list and CN a bit more deterministic
This is an edge case, but when a vhost has `https_ignore` present the CN
(and hence the base filename for the certificate and key) is selected
from the `dns_names` set using `pop` before the contents of the
`https_ignore` list are removed. This means that the CN used isn't chosen
in a deterministic fashion and there is a small risk that it may be in
the `https_ignore` list.

This change removes the contents of `https_ignore` before simply
returning the sorted contents of the `dns_names` set in cases where the
usual CN should be ignored. Behaviour in other cases should be unchanged.

This makes it easier to determine elsewhere what filenames to expect for
a given vhost as these are now generated in a slightly more predictable
manner.