letsencrypt: catch exceptions from call to simp_le
When trying to obtain a certificate `letsencrypt.py` calls out via
`subprocess.check_call` to `simp_le`. This commit adds some exception
handling so that any errors returned by this call are caught and
reported rather than halting execution.

This was an issue when calls were being made for certificate renewals
as a problem with one certificate resulted in execution halting and no
further renewals being processed. By catching the error and printing
an error statement, we should get a log of the problem but continue
to get all other certificates renewed.