acme-sh-helper: Use Cloudflare plugin for WDTK renewals
We've moved WDTK DNS behind Cloudflare for the time being. This is a bit
of a hack but should permit renewing the wildcard certs using their API.

We probably want a flag in config somewhere that will allow us to select
a particular provider, etc, but this will do for now.